Fixing buggy buildTree function
Using sets to track duplication in buildTree was creating an imbalanced tree.
As a solution, I created a sorted array without duplicates in the BST constructor
and used array.slice instead of using start and end indexes as arguments within the
buildTree function. This has resolved the issue ad created a balanced tree.